refactor: standardize Trash service file naming convention
