+++ src/app/shared/components/Breadcrumbs/BreadcrumbsMenu/BreadcrumbsMenuDrive.tsx
@@ -10,7 +10,7 @@ import {
 } from '@phosphor-icons/react';
 import MoveActionIcon from 'assets/icons/move.svg?react';
 import { useTranslationContext } from '../../../../i18n/provider/TranslationProvider';
-import moveItemsToTrash from '../../../../../views/Trash/services/move-items-to-trash';
+import moveItemsToTrash from '../../../../../views/Trash/services/move-items-to-trash.service';
 import { DriveItemData, DriveItemDetails, FolderPath } from '../../../../drive/types';
 import { useAppDispatch, useAppSelector } from '../../../../store/hooks';
 import { uiActions } from '../../../../store/slices/ui';
