+++ src/views/Trash/index.ts
@@ -4,7 +4,7 @@ export { default as DeleteItemsDialog } from './components/DeleteItemsDialog';
 export { default as StopSharingAndMoveToTrashDialogWrapper } from './components/StopSharingAndMoveToTrashDialogWrapper';
 export { default as StopSharingAndMoveToTrashDialog } from './components/StopSharingAndMoveToTrashDialog';
 export { useTrashPagination } from './hooks/useTrashPagination';
-export { getTrashPaginated, getWorkspaceTrashPaginated } from './services/get_trash';
-export { default as clearTrash } from './services/clear-trash';
-export { default as deleteItems } from './services/delete-items';
-export { default as moveItemsToTrash } from './services/move-items-to-trash';
+export { getTrashPaginated, getWorkspaceTrashPaginated } from './services/get-trash.service';
+export { default as clearTrash } from './services/clear-trash.service';
+export { default as deleteItems } from './services/delete-items.service';
+export { default as moveItemsToTrash } from './services/move-items-to-trash.service';
