+++ src/views/Trash/services/batch-processor.service.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, vi } from 'vitest';
-import { processBatchConcurrently } from './batch-processor';
+import { processBatchConcurrently } from './batch-processor.service';
 
 describe('processBatchConcurrently', () => {
   it('should process items in batches with concurrent limit', async () => {
@@ -17,21 +17,5 @@ describe('processBatchConcurrently', () => {
     expect(processor).toHaveBeenNthCalledWith(1, items.slice(0, 10));
     expect(processor).toHaveBeenNthCalledWith(2, items.slice(10, 20));
     expect(processor).toHaveBeenNthCalledWith(3, items.slice(20, 25));
-  });
-
-  it('should process when items equals max concurrent batches', async () => {
-    const items = Array.from({ length: 4 }, (_, i) => i);
-    const processor = vi.fn().mockResolvedValue(undefined);
-
-    await processBatchConcurrently({
-      items,
-      batchSize: 2,
-      maxConcurrentBatches: 2,
-      processor,
-    });
-
-    expect(processor).toHaveBeenCalledTimes(2);
-    expect(processor).toHaveBeenNthCalledWith(1, [0, 1]);
-    expect(processor).toHaveBeenNthCalledWith(2, [2, 3]);
   });
 });
