+++ src/views/Trash/services/batch-processor.service.ts
