+++ src/views/Trash/services/clear-trash.service.ts
